requirements.txt
hydra-core==1.3.2
matplotlib==3.10.3
numpy==2.2.6
omegaconf==2.3.0
pandas==2.3.0
scikit_learn==1.7.0
seaborn==0.13.2
torch==2.6.0
tqdm==4.67.1
wandb==0.20.1
folium==0.20.0
geopandas==1.1.1
skl2onnx==1.19.1
xgboost==3.0.5
shap==0.48.0
pyarrow
fastparquet